build: set up and document build and release scripts
Since we have a build step now we must point the package.json file
reference to the dist folder and must explicitly list all files that we
want to be included in the npm tarball.

Also instructions for relasing a new version have not yet been
documented which is finally done here.